+++ java/src/main/java/org/rocksdb/HistogramType.java
@@ -162,6 +162,12 @@ public enum HistogramType {
 
   ASYNC_READ_BYTES((byte) 0x33),
 
+  /**
+   * Number of bytes read in prefetching contents from the end of SST table
+   * during table open
+   */
+  TABLE_PREFETCH_TAIL_READ_BYTES((byte) 0x39),
+
   // 0x1F for backwards compatibility on current minor version.
   HISTOGRAM_ENUM_MAX((byte) 0x1F);